+++ Assets/Scripts/Networking/Opcodes/In/InServerMessage.cs
@@ -12,7 +12,7 @@ public class InServerMessage : OpcodeIn
     public override void Receive(MsgManager msgManager, ServerInput serverInput)
     {
         byte serverMessageType = msgManager.ReadByte();
-        byte length = msgManager.ReadByte();
+        ushort length = msgManager.ReadShort();
         serverInput.MessageHolder.Add(new System.Tuple<ServerMessageType, string>((ServerMessageType) serverMessageType, msgManager.ReadString(length)));
     }
 }
